In FileAPI, import and use GameList from resources subfolder.
